Move common functions to utility class
